Break out the "team matchups" calculation phase into a separate module that can be included/executed from another ruby file, or from the command line.  Handles odd # of teams (e.g. 7 teams in a 3-timeslot schedule) where one team has a bye each week.
Still using a questionable heuristic about whether to create a
single schedule block & repeat it, or to randomly come up with
different head-to-head pairings for each week block.  I think if
I do the timeslot scheduling better, I hope to be able to use
a repeating schedule and get well balanced calendars.  A schedule
with an odd number of teams cannot use a repeating schedule.