Update thursday schedule to have the correct timeslots -- different times for san mateo and for redwood city.
The analyzer script won't identify streaks of late games with two
different timeslots.  Will need to do that via attributes.

Added a new summary to the analyzer script to show the total
number/percentage of late and early games.  To make it easier
to spot these problems even if the timeslots are slightly different.