Fix prase-ics.rb and analyze-schedule.rb to correctly handle a schedule with a bye team (e.g. 9 teams playing in a 4-games-per-week schedule).
Change ice-oasis-leagues.rb to exit with an error result code instead
of success if there is an error.  Send error message to stderr.
Change Makefile to run this command before it removes the files
on the web server; if there is an internal issue we'll now stop
running the makefile before anything is removed from the server.