Finish up the ability to schedule a one of a league's game on a different day than the primary day.
For instance, a Thursday league might have 4 games on Thursday
and 1 game on Friday.  This can now be expressed via the :alternate_day
and :alternate_day_offset entries in the timeslots.

It should be possible to have a Thursday and Friday league combination
where the Thursday overflow timeslot alternates between 7pm and 10:45pm
timeslots every other week.  And Friday would similarly have a schedule
that shifts between 8:15, 9:30, 10:45 and 7:00, 8:15, 9:30 every other
week to accomodate.  This would be done with a list of game times in
the ice-oasis-leagues.rb that covers the timeslots for a two week block.

I haven't tested the full combination of this yet, but everything was
written to allow for this so it should be pretty close if it doesn't
work out of the gate.