Change create.rb to save the .ics files in /tmp.
Update create-ics-file.rb to emit a
X-APPLE-TRAVEL-ADVISORY-BEHAVIOR:DISABLED
entry which should prevent iphones from listing
travel time estimates for every game in the schedule.
We don't give people a way to thin the schedue down
to just their team so they would get alerts to all of
their night's games if they subscribed.

Also try emitting a X-HOCKEY-SCHEDULE-DEETS header which
will have the attributes for the timeslot so we can
do correct analysis of a schedule that came from an
.ics file.